requirements.txt
httpx>=0.27
python-telegram-bot>=21.0
anthropic>=0.40
python-dotenv>=1.0
eth-account>=0.13
web3>=7.0
